[BACK-53] Project Sentiment and Categorization (#32)
* ###Description: Implement description/ELI5 regarding the sentiment towards the project
###How I implemented: A file that contains all the chatGPT methods and a file that scrapes hackernews.
###Other I am currently working on the categorization wihtin the same branch. Sorry. But I am still struggleing with the types

In ScrapeRepoCategroies there is still a problem with the types (not eslint confirm) but since we need the functionality I will commit it know and modify it later. But I am just not able to to write the correct types for the complex response. There are prototypes within the types folder.

* Description of issue
Implement description/ ELI5 regarding the sentiment towards the project

How I implemented
A file that contains all the ChatGPT methods and a file that scrapes Hackernews. On more file that scrapes github.

Other
I am currently working on the categorization within the same branch. Sorry. But I am still struggling with the types.

In ScrapeRepoCategroies there is still a problem with the types (not eslint conform) but since we need the functionality I will commit it know and modify it later. But I am just not able to to write the correct types for the complex response. There are prototypes within the types folder.

* ###Description: Implement description/ELI5 regarding the sentiment towards the project
###How I implemented: A file that contains all the chatGPT methods and a file that scrapes hackernews.
###Other I am currently working on the categorization wihtin the same branch. Sorry. But I am still struggleing with the types

In ScrapeRepoCategroies there is still a problem with the types (not eslint confirm) but since we need the functionality I will commit it know and modify it later. But I am just not able to to write the correct types for the complex response. There are prototypes within the types folder.

* Delete openAIApi.ts

unwanted oush

* Delete chatGTPMethods.ts

* Delete hackerNewsScraper.ts

* Delete scrapeCategories.ts

* implement requested changes

* Delete scrapeGithubCategories.js

* with prettify

* ran prettify

* deleted hackernewstest

* changed var name

* changed parameters

* test

* changed parameters

* test

* Tom test

* test

* Test

* t

* test

* test

* Implemented BACK-64, by returning two values in serachHackernewsStories

* added wishes changes

* added wished changes

* I only wrote tests for my github methods. I hope that is ok. Can always add the other ones later

* test

* cant push currently

* test

* test

* test

* test

* test

* new test file

* changed test method. Still needs adjustment

* t

* t

* tests improved

* test

* Restore githubScraping.ts. Fix small issues

* Fix small issues

---------

Co-authored-by: Alexander Cornett <ge42nir@mytum.de>
Co-authored-by: Tom Woehrle <thomas.e.woehrle@gmail.com>
Co-authored-by: thomas-woehrle <127286604+thomas-woehrle@users.noreply.github.com>